RIDER-132995 refactoring
